Remove handler safe dlg
